refactor: rename package to match usual python convention
